[REFACT] Lecture 패키지 클래스 구조 정리
