+++ legend-momo-city/src/main/java/com/wanted/momocity/lecture/application/service/LectureCommandService.java
@@ -1,16 +1,31 @@
 package com.wanted.momocity.lecture.application.service;
 
+import com.wanted.momocity.global.application.s3.S3UploadPort;
 import com.wanted.momocity.global.domain.common.exception.DomainRuleViolationException;
-import com.wanted.momocity.lecture.application.command.ChangeLectureStatusCommand;
-import com.wanted.momocity.lecture.application.command.CreateLectureCommand;
+import com.wanted.momocity.lecture.application.command.LectureCommand.AdminChangeLectureStatusCommand;
+import com.wanted.momocity.lecture.application.command.LectureCommand.ChangeChapterVideoStatusCommand;
+import com.wanted.momocity.lecture.application.command.LectureCommand.ChangeLectureStatusCommand;
+import com.wanted.momocity.lecture.application.command.LectureCommand.CreateChapterCommand;
+import com.wanted.momocity.lecture.application.command.LectureCommand.CreateLectureCommand;
+import com.wanted.momocity.lecture.application.command.LectureCommand.RegisterChapterVideoCommand;
 import com.wanted.momocity.lecture.application.port.TeacherAccountPort;
-import com.wanted.momocity.lecture.application.usecase.LectureCommandUseCase;
+import com.wanted.momocity.lecture.application.usecase.LectureCommandUseCases.AdminLectureCommandUseCase;
+import com.wanted.momocity.lecture.application.usecase.LectureCommandUseCases.ChapterCommandUseCase;
+import com.wanted.momocity.lecture.application.usecase.LectureCommandUseCases.LectureCommandUseCase;
 import com.wanted.momocity.lecture.domain.event.LectureCreatedEvent;
+import com.wanted.momocity.lecture.domain.exception.ChapterLimitExceededException;
+import com.wanted.momocity.lecture.domain.exception.ChapterNotFoundException;
+import com.wanted.momocity.lecture.domain.exception.ChapterVideoAlreadyExistsException;
+import com.wanted.momocity.lecture.domain.exception.DuplicateChapterOrderException;
 import com.wanted.momocity.lecture.domain.exception.LectureNotFoundException;
 import com.wanted.momocity.lecture.domain.model.LectureAggregate;
+import com.wanted.momocity.lecture.domain.model.LectureChapter;
 import com.wanted.momocity.lecture.domain.model.LectureStatus;
+import com.wanted.momocity.lecture.domain.model.VideoStatus;
 import com.wanted.momocity.lecture.domain.repository.ChapterRepository;
 import com.wanted.momocity.lecture.domain.repository.LectureRepository;
+import com.wanted.momocity.lecture.presentation.api.response.AdminLectureResponse.AdminChangeLectureStatusResponse;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.security.access.AccessDeniedException;
@@ -19,37 +34,40 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.Instant;
 
-// LectureCommandService는 강의 등록 유스케이스의 실제 흐름을 담당
+/**
+ * Lecture 명령 기능을 처리하는 Application Service.
+ *
+ * 기존 LectureCommandService, ChapterCommandService,
+ * AdminLectureCommandService를 하나로 합친 형태.
+ */
 @Service
+@RequiredArgsConstructor
 @Transactional
 @Slf4j
-public class LectureCommandService implements LectureCommandUseCase {
+public class LectureCommandService implements
+        LectureCommandUseCase,
+        ChapterCommandUseCase,
+        AdminLectureCommandUseCase {
+
+    // 강의당 등록 가능한 최대 챕터 수
+    private static final int MAX_CHAPTER_COUNT = 5;
+
+    // 챕터 동영상 최대 업로드 크기: 500MB
+    private static final long MAX_VIDEO_SIZE_BYTES = 500 * 1024 * 1024;
 
     private final LectureRepository lectureRepository;
-    private final TeacherAccountPort teacherAccountPort;
     private final ChapterRepository chapterRepository;
+    private final TeacherAccountPort teacherAccountPort;
+    private final S3UploadPort s3UploadPort;
     private final ApplicationEventPublisher eventPublisher;
 
-    public LectureCommandService(
-            LectureRepository lectureRepository,
-            TeacherAccountPort teacherAccountPort,
-            ChapterRepository chapterRepository,
-            ApplicationEventPublisher eventPublisher
-    ) {
-        this.lectureRepository = lectureRepository;
-        this.teacherAccountPort = teacherAccountPort;
-        this.chapterRepository = chapterRepository;
-        this.eventPublisher = eventPublisher;
-    }
-
+    /**
+     * 강의 생성.
+     */
     @Override
     public LectureAggregate createLecture(CreateLectureCommand command) {
-        /*
-         * Authorization 토큰에서 얻은 email로 강사 id를 조회
-         */
         Long teacherId = teacherAccountPort.getTeacherId(command.teacherId());
 
-        // command.thumbnailUrl()은 S3 업로드 후 생성된 이미지 URL
         LectureAggregate lecture = LectureAggregate.create(
                 teacherId,
                 command.title(),
@@ -67,8 +85,6 @@ public class LectureCommandService implements LectureCommandUseCase {
                 Instant.now()
         ));
 
-        // 강의 등록이 정상 완료되었는지 추적하기 위한 로그
-        // S3 업로드 후 DB 저장까지 성공한 시점이므로 lectureId, teacherId, status만 남긴다.
         log.info("강의 등록 완료 - lectureId={}, teacherId={}, status={}",
                 savedLecture.getId(),
                 savedLecture.getTeacherId(),
@@ -77,46 +93,192 @@ public class LectureCommandService implements LectureCommandUseCase {
         return savedLecture;
     }
 
+    /**
+     * 강사가 본인 강의를 심사 대기 상태로 변경.
+     */
     @Override
     public LectureAggregate changeLectureStatus(ChangeLectureStatusCommand command) {
-        // Authorization 토큰에서 얻은 email로 강사 id를 조회
         Long teacherId = teacherAccountPort.getTeacherId(command.teacherId());
 
-        // 상태를 변경할 강의를 조회
         LectureAggregate lecture = lectureRepository.findById(command.lectureId())
                 .orElseThrow(() -> new LectureNotFoundException("강의를 찾을 수 없습니다."));
 
-        // 본인이 등록한 강의만 상태를 변경할 수 있음.
-        // 즉, 다른 강사가 로그인해서 해당 강사의 강인 승인 요청하는건 오류 뜸
         if (!lecture.isOwnedBy(teacherId)) {
             throw new AccessDeniedException("본인이 등록한 강의만 상태를 변경할 수 있습니다.");
         }
 
-        // 강사는 검수 요청 상태인 WAITING으로만 변경
         if (command.lectureStatus() != LectureStatus.WAITING) {
             throw new DomainRuleViolationException("강의 공개는 관리자만 할 수 있습니다.");
         }
 
-        /* comment
-         * WAITING은 강사가 강의를 등록 하자마자 대기 상태로 변한다.
-         * 관리자가 승인 해줘야 ACTIVE로 학생에게 공개
-         */
-        if (command.lectureStatus() == LectureStatus.WAITING) {
-            validateLectureReadyForReview(command.lectureId());
-        }
+        validateLectureReadyForReview(command.lectureId());
 
-        // 도메인 모델에 상태 변경을 요청
         LectureAggregate changedLecture = lecture.changeStatus(command.lectureStatus());
 
-        // 변경된 강의를 저장
         return lectureRepository.save(changedLecture);
     }
 
-    /* comment
-     * 강의를 ACTIVE 상태로 변경할 수 있는지 검증
-     * 조건:
-     * 1. 챕터가 최소 1개 이상 있어야 함
-     * 2. 모든 챕터에 동영상이 등록되어 있어야 함
+    /**
+     * 챕터 생성.
+     */
+    @Override
+    public LectureChapter createChapter(CreateChapterCommand command) {
+        Long teacherId = teacherAccountPort.getTeacherId(command.teacherId());
+
+        LectureAggregate lecture = lectureRepository.findById(command.lectureId())
+                .orElseThrow(() -> new LectureNotFoundException("강의를 찾을 수 없습니다."));
+
+        if (!lecture.isOwnedBy(teacherId)) {
+            throw new AccessDeniedException("본인이 등록한 강의에만 챕터를 등록할 수 있습니다.");
+        }
+
+        int chapterCount = chapterRepository.countByLectureId(command.lectureId());
+
+        if (chapterCount >= MAX_CHAPTER_COUNT) {
+            throw new ChapterLimitExceededException("챕터는 최대 5개까지만 등록할 수 있습니다.");
+        }
+
+        boolean duplicatedOrderNo = chapterRepository.existsByLectureIdAndOrderNo(
+                command.lectureId(),
+                command.orderNo()
+        );
+
+        if (duplicatedOrderNo) {
+            throw new DuplicateChapterOrderException("동일 강의 안에 이미 사용 중인 챕터 순서입니다.");
+        }
+
+        LectureChapter chapter = LectureChapter.create(
+                command.lectureId(),
+                command.title(),
+                command.orderNo()
+        );
+
+        LectureChapter savedChapter = chapterRepository.save(chapter);
+
+        log.info("챕터 등록 완료 - chapterId={}, lectureId={}, orderNo={}",
+                savedChapter.getId(),
+                savedChapter.getLectureId(),
+                savedChapter.getOrderNo());
+
+        return savedChapter;
+    }
+
+    /**
+     * 챕터 동영상 등록.
+     */
+    @Override
+    public LectureChapter registerChapterVideo(RegisterChapterVideoCommand command) {
+        Long teacherId = teacherAccountPort.getTeacherId(command.teacherId());
+
+        LectureAggregate lecture = lectureRepository.findById(command.lectureId())
+                .orElseThrow(() -> new LectureNotFoundException("강의를 찾을 수 없습니다."));
+
+        if (!lecture.isOwnedBy(teacherId)) {
+            throw new AccessDeniedException("본인이 등록한 강의의 챕터에만 동영상을 등록할 수 있습니다.");
+        }
+
+        LectureChapter chapter = chapterRepository.findById(command.chapterId())
+                .orElseThrow(() -> new ChapterNotFoundException("챕터를 찾을 수 없습니다."));
+
+        if (!chapter.belongsTo(command.lectureId())) {
+            throw new ChapterNotFoundException("유효하지 않은 챕터 식별자입니다.");
+        }
+
+        if (chapter.hasVideo()) {
+            throw new ChapterVideoAlreadyExistsException("이미 동영상이 등록된 챕터입니다.");
+        }
+
+        if (command.video() == null || command.video().isEmpty()) {
+            throw new DomainRuleViolationException("동영상 파일은 필수입니다.");
+        }
+
+        if (command.video().getSize() > MAX_VIDEO_SIZE_BYTES) {
+            throw new DomainRuleViolationException("동영상 파일 크기는 500MB 이하만 가능합니다.");
+        }
+
+        String videoUrl = s3UploadPort.upload(command.video());
+
+        LectureChapter updatedChapter = chapter.registerVideo(
+                videoUrl,
+                command.video().getSize(),
+                command.durationSec(),
+                command.video().getOriginalFilename()
+        );
+
+        LectureChapter savedChapter = chapterRepository.save(updatedChapter);
+
+        log.info("챕터 동영상 등록 완료 - chapterId={}, lectureId={}, videoStatus={}, videoSizeBytes={}",
+                savedChapter.getId(),
+                savedChapter.getLectureId(),
+                savedChapter.getVideoStatus(),
+                savedChapter.getVideoSizeBytes());
+
+        return savedChapter;
+    }
+
+    /**
+     * 챕터 동영상 처리 상태 변경.
+     */
+    @Override
+    public LectureChapter changeChapterVideoStatus(ChangeChapterVideoStatusCommand command) {
+        Long teacherId = teacherAccountPort.getTeacherId(command.teacherId());
+
+        LectureAggregate lecture = lectureRepository.findById(command.lectureId())
+                .orElseThrow(() -> new LectureNotFoundException("강의를 찾을 수 없습니다."));
+
+        if (!lecture.isOwnedBy(teacherId)) {
+            throw new AccessDeniedException("본인이 등록한 강의의 챕터만 동영상 상태를 변경할 수 있습니다.");
+        }
+
+        LectureChapter chapter = chapterRepository.findById(command.chapterId())
+                .orElseThrow(() -> new ChapterNotFoundException("챕터를 찾을 수 없습니다."));
+
+        if (!chapter.belongsTo(command.lectureId())) {
+            throw new ChapterNotFoundException("유효하지 않은 챕터 식별자입니다.");
+        }
+
+        if (!chapter.hasVideo()) {
+            throw new DomainRuleViolationException("동영상이 등록된 챕터만 상태를 변경할 수 있습니다.");
+        }
+
+        LectureChapter changedChapter = chapter.changeVideoStatus(command.videoStatus());
+
+        LectureChapter savedChapter = chapterRepository.save(changedChapter);
+
+        log.info("챕터 동영상 상태 변경 완료 - chapterId={}, lectureId={}, videoStatus={}",
+                savedChapter.getId(),
+                savedChapter.getLectureId(),
+                savedChapter.getVideoStatus());
+
+        return savedChapter;
+    }
+
+    /**
+     * 관리자가 강의를 승인 또는 거절 상태로 변경.
+     */
+    @Override
+    public AdminChangeLectureStatusResponse changeLectureStatus(AdminChangeLectureStatusCommand command) {
+        LectureAggregate lecture = lectureRepository.findById(command.lectureId())
+                .orElseThrow(() -> new LectureNotFoundException("강의를 찾을 수 없습니다."));
+
+        if (command.lectureStatus() == LectureStatus.ACTIVE) {
+            validateLectureReadyForApproval(command.lectureId());
+        }
+
+        LectureAggregate changedLecture = lecture.changeStatus(command.lectureStatus());
+
+        LectureAggregate savedLecture = lectureRepository.save(changedLecture);
+
+        log.info("관리자 강의 상태 변경 완료 - adminId={}, lectureId={}, lectureStatus={}",
+                command.adminId(),
+                savedLecture.getId(),
+                savedLecture.getStatus());
+
+        return AdminChangeLectureStatusResponse.from(savedLecture);
+    }
+
+    /**
+     * 강사가 강의를 심사 요청할 수 있는 상태인지 검증.
      */
     private void validateLectureReadyForReview(Long lectureId) {
         int chapterCount = chapterRepository.countByLectureId(lectureId);
@@ -130,6 +292,25 @@ public class LectureCommandService implements LectureCommandUseCase {
 
         if (hasChapterWithoutVideo) {
             throw new DomainRuleViolationException("강의 등록하려면 모든 챕터에 동영상이 등록되어야 합니다.");
+        }
+    }
+
+    /**
+     * 관리자가 강의를 승인할 수 있는 상태인지 검증.
+     */
+    private void validateLectureReadyForApproval(Long lectureId) {
+        int chapterCount = chapterRepository.countByLectureId(lectureId);
+
+        if (chapterCount < 1) {
+            throw new DomainRuleViolationException("강의를 승인하려면 최소 1개 이상의 챕터가 필요합니다.");
+        }
+
+        if (chapterRepository.existsByLectureIdAndVideoUrlIsNull(lectureId)) {
+            throw new DomainRuleViolationException("강의를 승인하려면 모든 챕터에 동영상이 등록되어야 합니다.");
+        }
+
+        if (chapterRepository.existsByLectureIdAndVideoStatusNot(lectureId, VideoStatus.READY)) {
+            throw new DomainRuleViolationException("강의를 승인하려면 모든 동영상이 READY 상태여야 합니다.");
         }
     }
 }
