+++ legend-momo-city/src/main/java/com/wanted/momocity/lecture/application/service/LectureQueryService.java
@@ -3,80 +3,78 @@ package com.wanted.momocity.lecture.application.service;
 import com.wanted.momocity.auth.application.port.LoadUserPort;
 import com.wanted.momocity.auth.domain.model.User;
 import com.wanted.momocity.enrollment.application.port.StudentAccountPort;
+import com.wanted.momocity.global.domain.common.exception.DomainRuleViolationException;
 import com.wanted.momocity.lecture.application.port.LectureEnrollmentQueryPort;
 import com.wanted.momocity.lecture.application.port.TeacherAccountPort;
-import com.wanted.momocity.lecture.application.query.GetLecturesQuery;
-import com.wanted.momocity.lecture.application.query.GetStudentLectureDetailQuery;
-import com.wanted.momocity.lecture.application.query.GetTeacherLectureDetailQuery;
-import com.wanted.momocity.lecture.application.query.GetTeacherLecturesQuery;
-import com.wanted.momocity.lecture.application.usecase.LectureQueryUseCase;
+import com.wanted.momocity.lecture.application.query.LectureQuery.GetAdminLectureDetailQuery;
+import com.wanted.momocity.lecture.application.query.LectureQuery.GetAdminLecturesQuery;
+import com.wanted.momocity.lecture.application.query.LectureQuery.GetLecturesQuery;
+import com.wanted.momocity.lecture.application.query.LectureQuery.GetStudentLectureDetailQuery;
+import com.wanted.momocity.lecture.application.query.LectureQuery.GetTeacherLectureDetailQuery;
+import com.wanted.momocity.lecture.application.query.LectureQuery.GetTeacherLecturesQuery;
+import com.wanted.momocity.lecture.application.usecase.LectureQueryUseCases.AdminLectureQueryUseCase;
+import com.wanted.momocity.lecture.application.usecase.LectureQueryUseCases.LectureQueryUseCase;
 import com.wanted.momocity.lecture.domain.exception.LectureNotFoundException;
 import com.wanted.momocity.lecture.domain.model.LectureAggregate;
 import com.wanted.momocity.lecture.domain.model.LectureChapter;
 import com.wanted.momocity.lecture.domain.model.LectureStatus;
 import com.wanted.momocity.lecture.domain.repository.ChapterRepository;
 import com.wanted.momocity.lecture.domain.repository.LectureRepository;
-import com.wanted.momocity.lecture.presentation.api.response.*;
-
+import com.wanted.momocity.lecture.presentation.api.response.AdminLectureResponse.AdminLectureDetailResponse;
+import com.wanted.momocity.lecture.presentation.api.response.AdminLectureResponse.AdminLectureListItemResponse;
+import com.wanted.momocity.lecture.presentation.api.response.AdminLectureResponse.AdminLecturePageResponse;
+import com.wanted.momocity.lecture.presentation.api.response.StudentLectureResponse.StudentLectureDetailResponse;
+import com.wanted.momocity.lecture.presentation.api.response.StudentLectureResponse.StudentLectureListItemResponse;
+import com.wanted.momocity.lecture.presentation.api.response.StudentLectureResponse.StudentLecturePageResponse;
+import com.wanted.momocity.lecture.presentation.api.response.TeacherLectureResponse.TeacherLectureDetailResponse;
+import com.wanted.momocity.lecture.presentation.api.response.TeacherLectureResponse.TeacherLectureListItemResponse;
+import com.wanted.momocity.lecture.presentation.api.response.TeacherLectureResponse.TeacherLecturePageResponse;
 import lombok.RequiredArgsConstructor;
-
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-
 import java.util.List;
 
-/* comment
- * LectureQueryService는 강의 목록 조회 로직을 처리하는 서비스
- * enrolled=true / false 조건에 따라
- * 로그인 사용자의 수강 신청 여부를 기준으로 강의를 필터링함.
+/**
+ * 강의 조회 기능을 처리하는 Application Service.
+ *
+ * 기존 LectureQueryService와 AdminLectureQueryService를 하나로 합친 형태.
+ * 학생, 강사, 관리자 강의 조회 기능을 모두 담당한다.
  */
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class LectureQueryService implements LectureQueryUseCase {
+public class LectureQueryService implements
+        LectureQueryUseCase,
+        AdminLectureQueryUseCase {
 
-    // 강의 데이터를 조회하는 저장소
+    // 강의 조회 저장소
     private final LectureRepository lectureRepository;
 
-    // 로그인 사용자 email로 userId를 찾기 위한 포트
-    private final StudentAccountPort studentAccountPort;
-
-    // 사용자의 수강 신청 정보를 조회하기 위한 포트
-    private final LectureEnrollmentQueryPort lectureEnrollmentQueryPort;
-
-    private final TeacherAccountPort teacherAccountPort;
-
-    // 챕터 목록 조회
+    // 챕터 조회 저장소
     private final ChapterRepository chapterRepository;
 
-    // 강사 이름과 프로필 이미지를 조회하기 위한 auth 포트
+    // 로그인한 사용자의 학생 ID를 조회하는 포트
+    private final StudentAccountPort studentAccountPort;
+
+    // 로그인한 사용자의 강사 ID를 조회하는 포트
+    private final TeacherAccountPort teacherAccountPort;
+
+    // 수강 신청 정보 조회 포트
+    private final LectureEnrollmentQueryPort lectureEnrollmentQueryPort;
+
+    // 강사 이름, 프로필 이미지 조회를 위한 auth 포트
     private final LoadUserPort loadUserPort;
 
+    // 학생 기준 강의 목록 조회.
     @Override
-    @Transactional(readOnly = true)
     public StudentLecturePageResponse getLectures(GetLecturesQuery query) {
-
-        /*
-         * Authorization 토큰에서 꺼낸 로그인 사용자 ID로 학생 ID를 조회한다.
-         * 컨트롤러에서는 인증된 사용자 ID만 넘기고,
-         * 실제 학생 식별자는 studentAccountPort를 통해 가져온다.
-         */
         Long studentId = studentAccountPort.getStudentId(query.userId());
 
-        /*
-         * 현재 학생이 이미 수강 신청한 강의 ID 목록을 조회한다.
-         * 이 목록은 isEnrolled 값을 만들고,
-         * enrolled=true/false 필터링에도 사용된다.
-         */
         List<Long> enrolledLectureIds =
                 lectureEnrollmentQueryPort.findLectureIdsByUserId(studentId);
 
-        /*
-         * 학생에게 보여줄 강의 목록을 조회한다.
-         * category, keyword, enrolled 조건과 페이지네이션은 repository에서 처리한다.
-         */
         var lecturePage = lectureRepository.findLectures(
                 query.category(),
                 query.keyword(),
@@ -86,22 +84,13 @@ public class LectureQueryService implements LectureQueryUseCase {
                 query.size()
         );
 
-        /*
-         * 조회된 강의들을 목록 응답 item으로 변환한다.
-         * 강의가 1개여도 content 배열 안에 들어가야 하므로 List로 변환한다.
-         */
         List<StudentLectureListItemResponse> content = lecturePage.content().stream()
-                .map(lecture -> toResponse(
+                .map(lecture -> toStudentListItemResponse(
                         lecture,
-                        enrolledLectureIds.contains(lecture.getId()),
-                        studentId
+                        enrolledLectureIds.contains(lecture.getId())
                 ))
                 .toList();
 
-        /*
-         * 최종 data 영역을 만든다.
-         * Controller에서는 이 객체를 ApiResponse.data에 넣어 응답한다.
-         */
         return new StudentLecturePageResponse(
                 content,
                 query.page(),
@@ -111,66 +100,44 @@ public class LectureQueryService implements LectureQueryUseCase {
         );
     }
 
-    // 학생 강의 상세 조회
+    // 학생 기준 강의 상세 조회.
     @Override
     public StudentLectureDetailResponse getStudentLectureDetail(GetStudentLectureDetailQuery query) {
+        Long studentId = studentAccountPort.getStudentId(query.userId());
 
-        // Authorization 토큰에서 가져온 userId가 실제 학생 계정인지 확인합니다.
-        Long userId = studentAccountPort.getStudentId(query.userId());
-
-        // lectureId로 강의를 조회합니다.
         LectureAggregate lecture = lectureRepository.findById(query.lectureId())
                 .orElseThrow(() -> new LectureNotFoundException("강의를 찾을 수 없습니다."));
 
-        // 학생은 ACTIVE 상태의 강의만 상세 조회할 수 있음
         if (lecture.getStatus() != LectureStatus.ACTIVE) {
             throw new AccessDeniedException("진행 중인 강의만 조회할 수 있습니다.");
         }
 
-        // 해당 강의의 챕터 목록을 orderNo 오름차순으로 조회
         List<LectureChapter> chapters =
                 chapterRepository.findAllByLectureIdOrderByOrderNoAsc(query.lectureId());
 
-        // 로그인한 학생이 이 강의를 수강신청했는지 확인
         boolean isEnrolled = lectureEnrollmentQueryPort
-                .findByUserIdAndLectureId(userId, query.lectureId())
+                .findByUserIdAndLectureId(studentId, query.lectureId())
                 .isPresent();
 
-        // lecture.teacherId로 강사 정보를 조회
         User teacher = loadUserPort.findById(lecture.getTeacherId())
                 .orElseThrow(() -> new LectureNotFoundException("강사 정보를 찾을 수 없습니다."));
 
-        // 강사 이름
-        String teacherName = teacher.getName();
-
-        // 강사 프로필 이미지 URL
-        String teacherProfileImageUrl = teacher.getProfileImageUrl();
-
-        // 리뷰 집계는 다음 단계에서 review 테이블 조회로 연결
-        double averageRating = 0.0;
-        int reviewCount = 0;
-
-        // 조회한 강의/챕터/부가정보를 학생 상세 응답 DTO로 변환
         return StudentLectureDetailResponse.from(
                 lecture,
                 chapters,
-                teacherName,
-                teacherProfileImageUrl,
-                averageRating,
-                reviewCount,
+                teacher.getName(),
+                teacher.getProfileImageUrl(),
+                0.0,
+                0,
                 isEnrolled
         );
     }
 
-    // 강사가 본인이 등록한 강의 목록을 조회
+    // 강사 기준 본인 강의 목록 조회.
     @Override
     public TeacherLecturePageResponse getTeacherLectures(GetTeacherLecturesQuery query) {
-
-        // Authorization 토큰에서 가져온 email로 강사 ID를 조회
         Long teacherId = teacherAccountPort.getTeacherId(query.teacherId());
 
-        // 강의 목록을 조회
-        // category, keyword, enrolled 조건은 repository에서 처리
         var lecturePage = lectureRepository.findTeacherLectures(
                 teacherId,
                 query.category(),
@@ -179,7 +146,6 @@ public class LectureQueryService implements LectureQueryUseCase {
                 query.size()
         );
 
-        // 도메인 모델을 강사용 목록 응답 DTO로 변환
         List<TeacherLectureListItemResponse> content = lecturePage.content().stream()
                 .map(lecture -> TeacherLectureListItemResponse.from(
                         lecture,
@@ -188,7 +154,6 @@ public class LectureQueryService implements LectureQueryUseCase {
                 ))
                 .toList();
 
-        // 페이지 정보와 목록 응답을 함께 반환
         return new TeacherLecturePageResponse(
                 content,
                 query.page(),
@@ -198,27 +163,21 @@ public class LectureQueryService implements LectureQueryUseCase {
         );
     }
 
-    // 강사가 본인이 등록한 강의 상세 정보와 챕터 목록을 조회
+    // 강사 기준 본인 강의 상세 조회.
     @Override
     public TeacherLectureDetailResponse getTeacherLectureDetail(GetTeacherLectureDetailQuery query) {
-
-        // 토큰에서 꺼낸 teacherId가 실제 강사 계정인지 확인
         Long teacherId = teacherAccountPort.getTeacherId(query.teacherId());
 
-        // lectureId로 강의를 조회
         LectureAggregate lecture = lectureRepository.findById(query.lectureId())
                 .orElseThrow(() -> new LectureNotFoundException("강의를 찾을 수 없습니다."));
 
-        // 본인이 등록한 강의가 아니면 조회할 수 없음
         if (!lecture.isOwnedBy(teacherId)) {
             throw new AccessDeniedException("본인이 등록한 강의만 조회할 수 있습니다.");
         }
 
-        // 해당 강의의 챕터 목록을 orderNo 오름차순으로 조회
         List<LectureChapter> chapters =
                 chapterRepository.findAllByLectureIdOrderByOrderNoAsc(query.lectureId());
 
-        // 강의 정보와 챕터 목록을 응답 DTO로 변환
         return TeacherLectureDetailResponse.from(
                 lecture,
                 chapters,
@@ -227,35 +186,92 @@ public class LectureQueryService implements LectureQueryUseCase {
         );
     }
 
-    // Lecture 도메인 객체를 학생 강의 목록 응답 DTO로 변환
-    private StudentLectureListItemResponse toResponse(
-            LectureAggregate lecture,
-            boolean enrolled,
-            Long userId
-    ) {
-        // 강사 이름 조회는 Null -> 나중에 검색 기능 확장 가능
-        String teacherName = null;
+    /* comment
+     * 관리자 기준 강의 목록 조회.
+     * status가 없으면 WAITING, ACTIVE 강의를 함께 조회한다.
+     */
+    @Override
+    public AdminLecturePageResponse getAdminLectures(GetAdminLecturesQuery query) {
+        List<LectureStatus> statuses = resolveAdminLectureStatuses(query.status());
 
-        // 강의평은 module 4 때 구현 할 예정 -> 현재는 0으로 둠
-        double averageRating = 0.0;
-        int reviewCount = 0;
+        var lecturePage = lectureRepository.findAdminLectures(
+                statuses,
+                query.category(),
+                query.keyword(),
+                query.page(),
+                query.size()
+        );
 
-        return new StudentLectureListItemResponse(
-                lecture.getId(),                  // lectureId: 강의 ID
-                lecture.getTeacherId(),           // teacherId: 강사 ID
-                teacherName,                      // teacherName: 강사 이름
-                lecture.getTitle(),               // title: 강의 제목
-                lecture.getDescription(),         // description: 강의 설명
-                lecture.getThumbnailUrl(),        // thumbnailUrl: 썸네일 URL
-                lecture.getCategory().name(),     // category: 강의 카테고리
-                lecture.getStatus().name(),       // lectureStatus: 강의 상태
-                lecture.getCompletedUserCount(),  // completedUserCount: 완료한 사용자 수
-                averageRating,                    // averageRating: 평균 평점
-                reviewCount,                      // reviewCount: 리뷰 개수
-                enrolled,                         // isEnrolled: 수강신청 여부
-                lecture.getCreatedAt()            // createdAt: 강의 등록일
+        List<AdminLectureListItemResponse> content = lecturePage.content().stream()
+                .map(lecture -> AdminLectureListItemResponse.from(
+                        lecture,
+                        0.0,
+                        0
+                ))
+                .toList();
+
+        return new AdminLecturePageResponse(
+                content,
+                query.page(),
+                query.size(),
+                lecturePage.totalElements(),
+                lecturePage.totalPages()
         );
     }
 
+    // 관리자 기준 강의 상세 조회.
+    @Override
+    public AdminLectureDetailResponse getAdminLectureDetail(GetAdminLectureDetailQuery query) {
+        LectureAggregate lecture = lectureRepository.findById(query.lectureId())
+                .orElseThrow(() -> new LectureNotFoundException("강의를 찾을 수 없습니다."));
 
+        if (lecture.getStatus() != LectureStatus.WAITING
+                && lecture.getStatus() != LectureStatus.ACTIVE) {
+            throw new DomainRuleViolationException("관리자는 승인 대기 또는 진행 중 강의만 조회할 수 있습니다.");
+        }
+
+        List<LectureChapter> chapters =
+                chapterRepository.findAllByLectureIdOrderByOrderNoAsc(query.lectureId());
+
+        return AdminLectureDetailResponse.from(
+                lecture,
+                chapters,
+                0.0,
+                0
+        );
+    }
+
+    // 관리자 목록 조회에서 사용할 강의 상태 목록을 결정
+    private List<LectureStatus> resolveAdminLectureStatuses(LectureStatus status) {
+        if (status == null) {
+            return List.of(
+                    LectureStatus.WAITING,
+                    LectureStatus.ACTIVE
+            );
+        }
+
+        return List.of(status);
+    }
+
+    // 학생 강의 목록용 응답 객체로 변환
+    private StudentLectureListItemResponse toStudentListItemResponse(
+            LectureAggregate lecture,
+            boolean enrolled
+    ) {
+        return new StudentLectureListItemResponse(
+                lecture.getId(),
+                lecture.getTeacherId(),
+                null,
+                lecture.getTitle(),
+                lecture.getDescription(),
+                lecture.getThumbnailUrl(),
+                lecture.getCategory().name(),
+                lecture.getStatus().name(),
+                lecture.getCompletedUserCount(),
+                0.0,
+                0,
+                enrolled,
+                lecture.getCreatedAt()
+        );
+    }
 }
